:recycle: 깃 그래프 컴포넌트를 분리한다
- 깃 그래프를 분리함에 따라 발생한 타입 선언은 GitGraph를 원천으로 한다.
- 따라서 experience interface는 gitgraph의 인터페이스를 상속받는 식으로 전개했다.
- 추가 사항으로, 리팩토링 정리 과정에서 필요 없는 주석들을 제거했다.
