+++ components/GitGraph/index.ts
@@ -1,0 +1,2 @@
+export { default as GitGraph } from './Main';
+export type { GitGraphExperienceInterface } from './types';
