+++ components/GitGraph/types.ts
@@ -1,0 +1,14 @@
+export interface GitGraphExperienceInterface {
+  title: string;
+  period: {
+    start: string;
+    end: string;
+  };
+  contents: string[];
+}
+
+export interface GitGraphInterface {
+  shouldDraw: boolean;
+  shouldShowHistory: boolean;
+  nowExperience: GitGraphExperienceInterface;
+}
