+++ tsconfig.json
@@ -28,6 +28,6 @@
       "~/*": ["./*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "components/GitGraph"],
   "exclude": ["node_modules"]
 }
